+++ sandbox/styleguide/views.py
@@ -1,9 +1,10 @@
 from django.conf import settings
+from django.contrib.auth.mixins import LoginRequiredMixin
 
 from py_css_styleguide.django.views import StyleguideViewMixin
 
 
-class StyleguideIndexView(StyleguideViewMixin):
+class StyleguideIndexView(LoginRequiredMixin, StyleguideViewMixin):
     """
     Display styleguide from CSS manifest
     """
@@ -13,3 +14,4 @@ class StyleguideIndexView(StyleguideViewMixin):
     manifest_json_filepath = settings.STYLEGUIDE_DUMP_PATH
     save_dump = settings.STYLEGUIDE_SAVE_DUMP
     development_mode = settings.DEBUG
+    raise_exception = True