+++ CHANGELOG.rst
@@ -3,6 +3,14 @@
 Changelog
 =========
 
+Development
+***********
+
+* Upgraded Diskette to v0.5.0 and update its configuration to ensure it is working
+  well and dumps are safe to load;
+* Made Styleguide view required to be authenticated;
+
+
 Version 0.4.1 - 2025/04/30
 **************************